backend/requirements.txt
fastapi
uvicorn
pandas
joblib
pymongo
bcrypt
python-dotenv
paho-mqtt
numpy
scikit-learn
imbalanced-learn
optuna
xgboost
lightgbm
shap
passlib
jose
jwt